+++ src/layout/components/Header.tsx
@@ -1,6 +1,6 @@
 import DynamicWrapper from '@shared/components/DynamicWrapper';
 import { ErrorModal } from '@shared/components/ErrorModal';
-import { headerWrapper, headerContents, headerLogo, headerRight } from '@layout/styles/header-style';
+import { headerWrapper, headerContents, headerLogo, headerRight } from '@layout/styles/headerStyle';
 import UserProfile from '@auth/components/UserProfile';
 import { useRouter } from 'next/router';
 import { ErrorBoundary } from 'react-error-boundary';
