+++ src/layout/styles/footerStyle.ts
